Remove validate option
We might bring this back sometime in the future, but no one has asked
for it and it's going to be very very rarely useful and only in large
teams where someone might move an old git tag to cover their tracks
about a mistake or some other malicious activity.

It just confuses people now.